Almaif: Fix a few regressions
Update the compile-API to the new way of returing an error code
instead of aborting.

Update the argbuffer alignment to MAX_EXTENDED_ALIGNMENT.

Manually replace the llvm memcpy intrinsic.

+++ lib/CL/devices/almaif/tce_builtins.cl
@@ -46,8 +46,8 @@ pocl_add_i32 (global const int *__restrict input1,
 }
 
 __kernel void
-pocl_abs_f32 (global const int *__restrict input,
-              global int *__restrict output)
+pocl_abs_f32 (global const float *__restrict input,
+              global float *__restrict output)
 {
   size_t i = get_global_id (0);
 #ifdef cl_TCE_ABSF
